Day 4, Part 2
The extra parsing was a total slog. I considered writing a simple rule system to keep the value validation more concise, but I'm over it. The main thing I wanted to achieve was to do this in a single pass, but only validate a passports values if the required keys were all present. This worked because I had to iterate all characters to find the keys, so in the process I store the values for the current passport only, in a string array that may get more detailed parsing later. I used a lot of built in integer parsing for the values. I'm not proud.